Refactor createWithClause to apply to PostgreServerExtension base
- Add overriding createWithClause to GreenplumServerExtension
- Add getTableModifiers to PostgreServerExtension and implementation to
PostgreServerExternsionBase.
[#162461208]
