+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/edit/PostgreTableManager.java
@@ -19,7 +19,6 @@ package org.jkiss.dbeaver.ext.postgresql.edit;
 import org.jkiss.code.NotNull;
 import org.jkiss.code.Nullable;
 import org.jkiss.dbeaver.DBException;
-import org.jkiss.dbeaver.ext.postgresql.PostgreUtils;
 import org.jkiss.dbeaver.ext.postgresql.model.*;
 import org.jkiss.dbeaver.model.DBConstants;
 import org.jkiss.dbeaver.model.DBPEvaluationContext;
@@ -31,7 +30,6 @@ import org.jkiss.dbeaver.model.impl.DBSObjectCache;
 import org.jkiss.dbeaver.model.impl.edit.SQLDatabasePersistAction;
 import org.jkiss.dbeaver.model.messages.ModelMessages;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
-import org.jkiss.utils.ArrayUtils;
 import org.jkiss.utils.CommonUtils;
 
 import java.util.List;
@@ -121,61 +119,7 @@ public class PostgreTableManager extends PostgreTableManagerBase implements DBEO
 
     @Override
     protected void appendTableModifiers(DBRProgressMonitor monitor, PostgreTableBase tableBase, NestedObjectCommand tableProps, StringBuilder ddl, boolean alter) {
-        if (tableBase instanceof PostgreTable) {
-            PostgreTable table = (PostgreTable) tableBase;
-            if (!alter) {
-                try {
-                    final List<PostgreTableInheritance> superTables = table.getSuperInheritance(monitor);
-                    if (!CommonUtils.isEmpty(superTables)) {
-                        ddl.append("\nINHERITS (");
-                        for (int i = 0; i < superTables.size(); i++) {
-                            if (i > 0) ddl.append(",");
-                            ddl.append(superTables.get(i).getAssociatedEntity().getFullyQualifiedName(DBPEvaluationContext.DDL));
-                        }
-                        ddl.append(")");
-                    }
-                } catch (DBException e) {
-                    log.error(e);
-                }
-            }
-        }
-
-        if (tableBase instanceof PostgreTableRegular) {
-            PostgreTableRegular table = (PostgreTableRegular) tableBase;
-            try {
-                if (!alter) {
-                    ddl.append(tableBase.getDataSource().getServerType().createWithClause(table, tableBase));
-                }
-                boolean hasOtherSpecs = false;
-                PostgreTablespace tablespace = table.getTablespace(monitor);
-                if (tablespace != null && table.isTablespaceSpecified()) {
-                    if (!alter) {
-                        ddl.append("\nTABLESPACE ").append(tablespace.getName());
-                    }
-                    hasOtherSpecs = true;
-                }
-                if (!alter && hasOtherSpecs) {
-                    ddl.append("\n");
-                }
-            } catch (DBException e) {
-                log.error(e);
-            }
-        } else if (tableBase instanceof PostgreTableForeign) {
-            PostgreTableForeign table = (PostgreTableForeign)tableBase;
-            try {
-                PostgreForeignServer foreignServer = table.getForeignServer(monitor);
-                if (foreignServer != null ) {
-                    ddl.append("\nSERVER ").append(DBUtils.getQuotedIdentifier(foreignServer));
-                }
-                String[] foreignOptions = table.getForeignOptions(monitor);
-                if (!ArrayUtils.isEmpty(foreignOptions)) {
-                    ddl.append("\nOPTIONS ").append(PostgreUtils.getOptionsString(foreignOptions));
-                }
-            } catch (DBException e) {
-                log.error(e);
-            }
-        }
-        tableBase.appendTableModifiers(monitor, ddl);
+        ddl.append(tableBase.getDataSource().getServerType().getTableModifiers(monitor, tableBase, alter));
     }
 
     @Override
