+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/PostgreServerExtension.java
@@ -78,4 +78,6 @@ public interface PostgreServerExtension
     void configureDialect(PostgreDialect dialect);
 
     String createWithClause(PostgreTableRegular table, PostgreTableBase tableBase);
+
+    String getTableModifiers(DBRProgressMonitor monitor, PostgreTableBase tableBase, boolean alter);
 }
