+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/impls/PostgreServerTimescale.java
@@ -20,8 +20,6 @@ import org.jkiss.dbeaver.ext.postgresql.model.PostgreDataSource;
 import org.jkiss.dbeaver.ext.postgresql.model.PostgreTableBase;
 import org.jkiss.dbeaver.ext.postgresql.model.PostgreTableRegular;
 
-import static org.jkiss.dbeaver.ext.postgresql.model.PostgreWithClauseBuilder.generateWithClause;
-
 /**
  * PostgreServerGreenplum
  */
@@ -35,11 +33,5 @@ public class PostgreServerTimescale extends PostgreServerExtensionBase {
     public String getServerTypeName() {
         return "Timescale";
     }
-
-    @Override
-    public String createWithClause(PostgreTableRegular table, PostgreTableBase tableBase) {
-        return generateWithClause(table);
-    }
-
 }
