+++ plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/impls/yellowbrick/PostgreServerYellowBrick.java
@@ -24,8 +24,6 @@ import org.jkiss.dbeaver.ext.postgresql.model.impls.PostgreServerExtensionBase;
 import org.jkiss.dbeaver.model.runtime.DBRProgressMonitor;
 import org.jkiss.utils.CommonUtils;
 
-import static org.jkiss.dbeaver.ext.postgresql.model.PostgreWithClauseBuilder.generateWithClause;
-
 /**
  * PostgreServerYellowBrick
  */
@@ -52,11 +50,5 @@ public class PostgreServerYellowBrick extends PostgreServerExtensionBase {
         }
         return null;
     }
-
-    @Override
-    public String createWithClause(PostgreTableRegular table, PostgreTableBase tableBase) {
-        return generateWithClause(table);
-    }
-
 }
